+++ lib/browser/base.rb
@@ -187,6 +187,11 @@ module Browser
       ua =~ /Opera Mini/ && detect_version?(full_version, expected_version)
     end
 
+    # Detect if browser is DuckDuckGo.
+    def duck_duck_go?(expected_version = nil)
+      ua =~ /DuckDuckGo/ && detect_version?(full_version, expected_version)
+    end
+
     def webkit_full_version
       ua[%r{AppleWebKit/([\d.]+)}, 1] || "0.0"
     end